+++ src/traits/HasAuthentication.php
@@ -10,12 +10,12 @@ trait HasAuthentication
   /**
    * Authenticator
    */
-  private AuthInterface|null $authenticator = null;
+  private Authenticator|null $authenticator = null;
 
   /**
    * Set authenticator
    */
-  public function authenticate(AuthInterface $authenticator): static
+  public function authenticate(Authenticator $authenticator): static
   {
     $this->_assertNotLocked('set authenticator');
     $this->authenticator = $authenticator;
@@ -27,7 +27,7 @@ trait HasAuthentication
    *
    * @internal
    */
-  public function _getAuthenticator(): AuthInterface|null
+  public function _getAuthenticator(): Authenticator|null
   {
     return $this->authenticator;
   }